+++ mod-source-record-manager-server/src/main/java/org/folio/dao/JournalRecordDaoImpl.java
@@ -2,6 +2,7 @@ package org.folio.dao;
 
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
+import io.vertx.core.json.JsonArray;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.RowSet;
 import io.vertx.sqlclient.Tuple;
@@ -36,6 +37,7 @@ import java.util.UUID;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
@@ -127,6 +129,20 @@ public class JournalRecordDaoImpl implements JournalRecordDao {
     }
     return promise.future().map(journalRecord.getId())
       .onFailure(e -> LOGGER.error("Error saving JournalRecord entity", e));
+  }
+
+  @Override
+  public Future<List<RowSet<Row>>> saveBatch(List<JournalRecord> journalRecords, String tenantId) {
+    Promise<List<RowSet<Row>>> promise = Promise.promise();
+    try {
+      List<Tuple> tupleList = journalRecords.stream().map(this::prepareInsertQueryParameters).collect(Collectors.toList());
+      String query = format(INSERT_SQL, convertToPsqlStandard(tenantId), JOURNAL_RECORDS_TABLE);
+      pgClientFactory.createInstance(tenantId).execute(query, tupleList, promise);
+    } catch (Exception e) {
+      LOGGER.error("Error saving JournalRecord entities", e);
+      promise.fail(e);
+    }
+    return promise.future().onFailure(e -> LOGGER.error("Error saving JournalRecord entities", e));
   }
 
   private Tuple prepareInsertQueryParameters(JournalRecord journalRecord) {